tests/user: Move tests into `routes` submodules
