+++ src/tests/routes/me/mod.rs
@@ -1,1 +1,4 @@
+mod email_notifications;
+pub mod get;
 pub mod tokens;
+mod updates;
