+++ src/tests/routes/session/mod.rs
@@ -1,0 +1,2 @@
+mod authorize;
+mod begin;
