+++ src/tests/routes/users/mod.rs
@@ -1,0 +1,3 @@
+mod read;
+mod stats;
+pub mod update;
